fix(coreml-export): extract guards to coremltools-free helpers module
Closes #3.

CI test collection has been failing since the Session 4 merge
because tests/test_coreml_export.py imports from
terncore.coreml_export, which imports coremltools at module top.
CI installs pip install -e ".[dev,transformers]" — coremltools
is in neither extra. Local venv masked the gap.

Shape B fix per the saved diagnosis memory: extract the two FP16
cast guards to a new numpy-only module. Tests import from helpers
directly, no transitive coremltools pull.

Changes:
- New: src/terncore/coreml_export_helpers.py (numpy-only)
  Carries FP16_MAX = 65504.0, _validate_ternary2_alpha,
  _cast_fp16_retain_with_guards. Verbatim move from
  coreml_export.py with explanatory module docstring.

- Modified: src/terncore/coreml_export.py
  Removes the constant and two helper definitions; adds import
  from terncore.coreml_export_helpers. Call sites in
  _load_weight_for_coreml unchanged (helper names resolve from
  the new module).

- Modified: tests/test_coreml_export.py
  Single-line import swap: terncore.coreml_export →
  terncore.coreml_export_helpers. Test bodies unchanged.

Validation:
- pytest tests/test_coreml_export.py: 7 passed (Session 4 baseline)
- pytest tests/ full suite: 310 passed, 19 skipped (no regressions)
- CI simulation (coremltools blocked via meta_path): import succeeds
- DSR1-7B export+palettise pipeline: still clean post-refactor

Pre-existing tests/test_stage1b.py::TestAccelInfo failures
(red since 2026-03-27) are tracked separately at #4 and not
addressed in this fix.
